+++ internal/js_lexer/js_lexer.go
@@ -223,6 +223,7 @@ type Lexer struct {
 	end                             int
 	ApproximateNewlineCount         int
 	LegacyOctalLoc                  logger.Loc
+	PreviousBackslashQuoteInJSX     logger.Range
 	Token                           T
 	HasNewlineBefore                bool
 	HasPureCommentBefore            bool
@@ -890,6 +891,7 @@ func (lexer *Lexer) NextInsideJSXElement() {
 			}
 
 		case '\'', '"':
+			var backslash logger.Range
 			quote := lexer.codePoint
 			needsDecode := false
 			lexer.step()
@@ -904,7 +906,16 @@ func (lexer *Lexer) NextInsideJSXElement() {
 					needsDecode = true
 					lexer.step()
 
+				case '\\':
+					backslash = logger.Range{Loc: logger.Loc{Start: int32(lexer.end)}, Len: 1}
+					lexer.step()
+					continue
+
 				case quote:
+					if backslash.Len > 0 {
+						backslash.Len++
+						lexer.PreviousBackslashQuoteInJSX = backslash
+					}
 					lexer.step()
 					break stringLiteral
 
@@ -915,6 +926,7 @@ func (lexer *Lexer) NextInsideJSXElement() {
 					}
 					lexer.step()
 				}
+				backslash = logger.Range{}
 			}
 
 			lexer.Token = TStringLiteral